Add LibGroup template
